// src/stores/types.ts
export interface FilmInfo {
    director:string
    name: string
    year: number
    gender: string []
    filmImage: string
    filmId: number
    cast: string[]
    valueRating: number
    runtimeMinutes: number
}

export interface SeriesInfo {
    director:string
    name: string
    year: number
    gender: string []
    filmImage: string
    seriesId: number
    cast: string[]
    valueRating: number
    runtimeMinutes: number
}